Pantalla Game Over Completa

+++ Assets/_CrystalGuardians/Estructuras/Extractor_obsidium/ExtractorObsidium.cs
@@ -49,7 +49,7 @@ public GameObject[] levels;
     private void generarRecursos()
     {
         GameManager.Instance.Obsiidum = GameManager.Instance.Obsiidum + generacionObsidiumPorNivel[nivelActual] * Time.deltaTime;
-        GameManager.Instance.ObsidiumTotalGenerado = GameManager.Instance.ObsidiumTotalGenerado +  (int)(generacionObsidiumPorNivel[nivelActual] * Time.deltaTime);
+        GameManager.Instance.ObsidiumTotalGenerado = GameManager.Instance.ObsidiumTotalGenerado + generacionObsidiumPorNivel[nivelActual] * Time.deltaTime;
 
     }
